third_party/blink/web_tests/fast/dom/shadow/adopt-node-with-shadow-root.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<div id="console"></div>
<object data="resources/svg_placeholder.svg" id="root" onload="test()"/>
<script>
function test() {
    var foreignDocument = document.getElementById('root').contentDocument;
    shadow = foreignDocument.getElementById('s').createShadowRoot();
    shouldThrow("document.adoptNode(shadow)");
}
</script>
</body>
</html>
